404 pages rendreed within layout
Before this pull request, the 404 and 404-search templates where handled
as full HTML pages without the page context (layout).

With this pull request both pages are rendered inside the selected
themes layout. Additionally the 404 template now contains the "search
form" and the homepage link.

All Themes are already updated with the above changes.