Update Font Awesome version in Admin panel
The current Font Awesome version in the admin panel is 4.4.0.

I updated this to the latest free version 6.6.0 here https://github.com/FortAwesome/Font-Awesome

I checked and all icons look to be loading properly in the admin panel. I also checked each of the built-in themes to make sure nothing was conflicting, and I didn't notice any missing icons in the themes.